+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [1.3.0] - 2024-01-14
 ### Added
+- Optional `--outputDir` flag
+- `outputDir` option in the config file
 
 ## [1.2.1] - 2024-01-13
 ### Changed